chore(custom-msg): Renamed files
